+++ src/utils/utftoupper.h
@@ -1,0 +1,18 @@
+// Copyright (c) 2026 Mikhail Kurinnoi
+// Distributed under the MIT License.
+// See the LICENSE file in the project root for more information.
+
+#ifndef UTILS_UTFTOUPPER_H
+#define UTILS_UTFTOUPPER_H
+
+#include <string>
+
+namespace dncdbg
+{
+
+// Convert UTF-8 string to uppercase.
+std::string to_uppercase(const std::string &input);
+
+} // namespace dncdbg
+
+#endif // UTILS_UTFTOUPPER_H