+++ CHANGELOG.md
@@ -33,6 +33,7 @@ Upcoming changes compared to previous version.
 - Replaced managed Roslyn parser with native tree-sitter implementation.
 - Implemented native unary operators for managed primitive types.
 - Implemented native binary operators for managed primitive types.
+- Implemented native UTF-8 string to uppercase conversion.
 
 #### Removed
 - Removed unused code.